+++ examples/nextjs-cjs/.eslintrc.js
@@ -1,0 +1,38 @@
+/**
+ * Specific eslint rules for this workspace, learn how to compose
+ * @link https://github.com/belgattitude/perso/tree/main/packages/eslint-config-bases
+ */
+
+// Workaround for https://github.com/eslint/eslint/issues/3458
+require('@soluble/eslint-config-bases/patch/modern-module-resolution');
+
+const {
+  getDefaultIgnorePatterns,
+} = require('@soluble/eslint-config-bases/helpers');
+
+module.exports = {
+  root: true,
+  parserOptions: {
+    tsconfigRootDir: __dirname,
+    project: 'tsconfig.json',
+  },
+  ignorePatterns: [...getDefaultIgnorePatterns()],
+  extends: [
+    '@soluble/eslint-config-bases/typescript',
+    '@soluble/eslint-config-bases/sonar',
+    '@soluble/eslint-config-bases/regexp',
+    '@soluble/eslint-config-bases/jest',
+    '@soluble/eslint-config-bases/react',
+    '@soluble/eslint-config-bases/rtl',
+    // Add specific rules for nextjs
+    'next/core-web-vitals',
+    // Apply prettier and disable incompatible rules
+    '@soluble/eslint-config-bases/prettier',
+  ],
+  rules: {
+    // optional overrides per project
+  },
+  overrides: [
+    // optional overrides per project file match
+  ],
+};