+++ packages/cache-ioredis/jest.config.js
@@ -24,9 +24,9 @@ const getTsConfigBasePaths = (tsConfigFile) => {
 
 /** @type {import('ts-jest/dist').InitialOptionsTsJest} */
 const config = {
-  displayName: `cache-interop:unit`,
+  displayName: `cache-ioredis:unit`,
   preset: 'ts-jest/presets/default-esm',
-  cacheDirectory: getJestCachePath('@soluble/cache-interop'),
+  cacheDirectory: getJestCachePath('@soluble/cache-ioredis'),
   testEnvironment: 'node',
   extensionsToTreatAsEsm: ['.ts'],
   verbose: true,