+++ packages/cache-interop/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @soluble/cache-interop
 
+## 0.11.1
+
+### Patch Changes
+
+- [#665](https://github.com/soluble-io/cache-interop/pull/665) [`40e9749`](https://github.com/soluble-io/cache-interop/commit/40e9749542bd81ac2c064384342e769e38e2d7ca) Thanks [@belgattitude](https://github.com/belgattitude)! - Fix dual publishing (commonjs, esm) with some tools
+
+* [#660](https://github.com/soluble-io/cache-interop/pull/660) [`6783b32`](https://github.com/soluble-io/cache-interop/commit/6783b3217e4caebfe0117ff9190bb3f54c014859) Thanks [@belgattitude](https://github.com/belgattitude)! - Improve linter
+
+## 0.11.1-canary.1
+
+### Patch Changes
+
+- [#665](https://github.com/soluble-io/cache-interop/pull/665) [`40e9749`](https://github.com/soluble-io/cache-interop/commit/40e9749542bd81ac2c064384342e769e38e2d7ca) Thanks [@belgattitude](https://github.com/belgattitude)! - Fix dual publishing (commonjs, esm) with some tools
+
+## 0.11.1-canary.0
+
+### Patch Changes
+
+- [#660](https://github.com/soluble-io/cache-interop/pull/660) [`6783b32`](https://github.com/soluble-io/cache-interop/commit/6783b3217e4caebfe0117ff9190bb3f54c014859) Thanks [@belgattitude](https://github.com/belgattitude)! - Improve linter
+
 ## 0.11.0
 
 ### Minor Changes