+++ packages/cache-ioredis/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @soluble/cache-ioredis
 
+## 0.12.1
+
+### Patch Changes
+
+- [#665](https://github.com/soluble-io/cache-interop/pull/665) [`40e9749`](https://github.com/soluble-io/cache-interop/commit/40e9749542bd81ac2c064384342e769e38e2d7ca) Thanks [@belgattitude](https://github.com/belgattitude)! - Fix dual publishing (commonjs, esm) with some tools
+
+* [#660](https://github.com/soluble-io/cache-interop/pull/660) [`6783b32`](https://github.com/soluble-io/cache-interop/commit/6783b3217e4caebfe0117ff9190bb3f54c014859) Thanks [@belgattitude](https://github.com/belgattitude)! - Improve linter
+
+* Updated dependencies [[`40e9749`](https://github.com/soluble-io/cache-interop/commit/40e9749542bd81ac2c064384342e769e38e2d7ca), [`6783b32`](https://github.com/soluble-io/cache-interop/commit/6783b3217e4caebfe0117ff9190bb3f54c014859), [`50a655d`](https://github.com/soluble-io/cache-interop/commit/50a655dc4db0eab0b63f0b73e14e0a998768d954)]:
+  - @soluble/cache-interop@0.11.1
+  - @soluble/dsn-parser@1.6.0
+
+## 0.12.1-canary.2
+
+### Patch Changes
+
+- [#665](https://github.com/soluble-io/cache-interop/pull/665) [`40e9749`](https://github.com/soluble-io/cache-interop/commit/40e9749542bd81ac2c064384342e769e38e2d7ca) Thanks [@belgattitude](https://github.com/belgattitude)! - Fix dual publishing (commonjs, esm) with some tools
+
+- Updated dependencies [[`40e9749`](https://github.com/soluble-io/cache-interop/commit/40e9749542bd81ac2c064384342e769e38e2d7ca)]:
+  - @soluble/cache-interop@0.11.1-canary.1
+  - @soluble/dsn-parser@1.6.0-canary.2
+
+## 0.12.1-canary.1
+
+### Patch Changes
+
+- Updated dependencies [[`50a655d`](https://github.com/soluble-io/cache-interop/commit/50a655dc4db0eab0b63f0b73e14e0a998768d954)]:
+  - @soluble/dsn-parser@1.6.0-canary.1
+
+## 0.12.1-canary.0
+
+### Patch Changes
+
+- [#660](https://github.com/soluble-io/cache-interop/pull/660) [`6783b32`](https://github.com/soluble-io/cache-interop/commit/6783b3217e4caebfe0117ff9190bb3f54c014859) Thanks [@belgattitude](https://github.com/belgattitude)! - Improve linter
+
+- Updated dependencies [[`6783b32`](https://github.com/soluble-io/cache-interop/commit/6783b3217e4caebfe0117ff9190bb3f54c014859)]:
+  - @soluble/cache-interop@0.11.1-canary.0
+  - @soluble/dsn-parser@1.5.1-canary.0
+
 ## 0.12.0
 
 ### Minor Changes